[RFC] frontend_reverse: Use extended attributes to cache file ID and Block0IV.
The usage of deterministic initialization vectors in reverse mode can lead to
various disadvantages compared to forward mode, in particular:

1) Renaming or moving a file causes a full retransmission due to the changed
   initialization vectors.

2) Hard-link handling is somewhat broken: Using the path no longer is really
   deterministic, and could lead to the usage of a different IV on each
   restart of Gocryptfs (depending on which path is accessed first). Current
   versions also never clear the cache at runtime, possibly leading to IV
   reusage for completely unrelated files when inode numbers are reused.

3) Attackers can detect when a block of a file is reverted to a previous state
   (but they still don't know the content).

An earlier version of this patchset, still available at [1], used a separate
database file linking inode numbers to initialization vectors (for each block).
While this solves all of the issues above, it increases disk space usage by up
to 0.8% (assuming 16 byte blocks + 16 byte checksums). Also, it adds a fairly
large amount of code to ensure the database remains in a consistent state.

Since we mostly care about points 1 and 2, and less about the third one, this
alternative implementation uses extended attributes to store the file ID and
Block0IV. Due to the size limitations of extended attributes, we can no longer
have truly random IVs for each block. Also, setting IVs means that gocryptfs
has to run as a user with sufficient privileges. The main advantage is that
this approach is much easier and straightforward, and no longer requires
maintaining a separate database.

For more details, see [2].

[1] https://github.com/slackner/gocryptfs/commits/nondet-reverse
[2] https://github.com/rfjakob/gocryptfs/issues/158